requirements.txt
# 虚拟机环境缺kenchi库
kenchi==0.10.0(pip install kenchi)

# 虚拟机环境 自带的
scikit-learn==0.21.3
scikit-image==0.13.1
keras==2.2.0
tensorflow-gpu==1.12.0
opencv-python==4.1.0.25
openslide-python==1.1.1
numpy==1.14.3
pandas==0.23.0
matplotlib==2.2.2
Pillow==5.1.0


